The time of warnings is over.

+++ include/SCD/CD/CD_Simplex.hxx
@@ -222,9 +222,9 @@ inline void CD_Simplex::filter(const CD_SimplexKeptPoints &f)
 			char a[]={0,1,2,3};
 			char b;
 			CD_SimplexExchangeTest<Vector3>(s1_,(*this)[f.b1],cache);
-			CD_SimplexExchange<char>(a[0],a[f.b1],b);
+			CD_SimplexExchange<char>(a[0],a[(int)f.b1],b);
 
-			s2_=(*this)[a[f.b2]];
+			s2_=(*this)[a[(int)f.b2]];
 			return;
 		}
 	case CD_Triangle:
@@ -233,13 +233,13 @@ inline void CD_Simplex::filter(const CD_SimplexKeptPoints &f)
 			Vector3 cache;
 			char a[]={0,1,2,3};
 			char b;
-			CD_SimplexExchangeTest<Vector3>(s1_,(*this)[f.b1],cache);
-			CD_SimplexExchange<char>(a[0],a[f.b1],b);
+			CD_SimplexExchangeTest<Vector3>(s1_,(*this)[(int)f.b1],cache);
+			CD_SimplexExchange<char>(a[0],a[(int)f.b1],b);
 
-			CD_SimplexExchangeTest<Vector3>(s2_,(*this)[a[f.b2]],cache);
-			CD_SimplexExchange<char>(a[a[1]],a[a[f.b2]],b);
+			CD_SimplexExchangeTest<Vector3>(s2_,(*this)[a[(int)f.b2]],cache);
+			CD_SimplexExchange<char>(a[(int)a[1]],a[(int)a[(int)f.b2]],b);
 
-			s3_=(*this)[a[f.b3]];
+			s3_=(*this)[a[(int)f.b3]];
 			return;
 			
 		}
@@ -247,7 +247,7 @@ inline void CD_Simplex::filter(const CD_SimplexKeptPoints &f)
 		{
 			Vector3 cache;
 			type_=CD_Point;
-			s1_=(*this)[f.b1];
+			s1_=(*this)[(int)f.b1];
 			return;
 		}
 	default:
@@ -256,16 +256,16 @@ inline void CD_Simplex::filter(const CD_SimplexKeptPoints &f)
 			Vector3 cache;
 			char a[]={0,1,2,3};
 			char b;
-			CD_SimplexExchangeTest<Vector3>(s1_,(*this)[f.b1],cache);
-			CD_SimplexExchange<char>(a[0],a[f.b1],b);
+			CD_SimplexExchangeTest<Vector3>(s1_,(*this)[(int)f.b1],cache);
+			CD_SimplexExchange<char>(a[0],a[(int)f.b1],b);
 
-			CD_SimplexExchangeTest<Vector3>(s2_,(*this)[a[f.b2]],cache);
-			CD_SimplexExchange<char>(a[a[1]],a[a[f.b2]],b);
+			CD_SimplexExchangeTest<Vector3>(s2_,(*this)[a[(int)f.b2]],cache);
+			CD_SimplexExchange<char>(a[(int)a[1]],a[(int)a[(int)f.b2]],b);
 
-			CD_SimplexExchangeTest<Vector3>(s3_,(*this)[a[f.b3]],cache);
-			CD_SimplexExchange<char>(a[a[2]],a[a[f.b3]],b);
+			CD_SimplexExchangeTest<Vector3>(s3_,(*this)[a[(int)f.b3]],cache);
+			CD_SimplexExchange<char>(a[(int)a[2]],a[(int)a[(int)f.b3]],b);
 
-			s4_=(*this)[a[f.b4]];
+			s4_=(*this)[a[(int)f.b4]];
 			return ;
 
 		}
